MultifileArrays

Documentation for MultifileArrays.

MultifileArrays.load_series

MultifileArrays.load_series

MultifileArrays.select_series

MultifileArrays.load_series — Method

A = load_series(f, filepattern; dir=pwd())

Create a lazily-loaded array A from a set of files. f(filename) should create an array from the filename, and filepattern is a pattern matching the names of the desired files. The file names should have a numeric portion that indicates ordering; ordering is numeric rather than alphabetical, so left-padding with zeros is optional.

Examples

Suppose you are currently in a directory with files image01.tiff ... image12.tiff. Then either

julia> using FileIO, MultifileArrays

julia> img = load_series(load, "image*.tiff")

or the more precise regular-expression form

julia> img = load_series(load, r"image(\d+).tiff");

suffice to load the image files.

MultifileArrays.load_series — Method

A = load_series(f, filenames::AbstractArray{<:AbstractString}, buffer::AbstractArray)

Create a lazily-loaded array A from a set of files. f is a function to load the data from a specific file into an array equivalent to buffer, meaning that

f(buffer, filename)

should fill buffer with the contents of filename.

filenames should be an array of file names with shape equivalent to the trailing dimensions of A, i.e., those that follow the dimensions of buffer.

The advantage of this syntax is that it provides greater control than load_series(f, filepattern) over the choice of files and the shape of the overall output.

Examples

Suppose you are currently in a directory with files image_z=1_t=1.tiff through image_z=5_t=30.tiff, where each file corresponds to a 2d (x, y) slice and the filename indicates the z and t coordinates. You could reshape filenames into matrix form

5×30 Matrix{String}:
 "image_z=1_t=1.tiff"  "image_z=1_t=2.tiff"  "image_z=1_t=3.tiff"  …  "image_z=1_t=29.tiff"  "image_z=1_t=30.tiff"
 "image_z=2_t=1.tiff"  "image_z=2_t=2.tiff"  "image_z=2_t=3.tiff"     "image_z=2_t=29.tiff"  "image_z=2_t=30.tiff"
 "image_z=3_t=1.tiff"  "image_z=3_t=2.tiff"  "image_z=3_t=3.tiff"     "image_z=3_t=29.tiff"  "image_z=3_t=30.tiff"
 "image_z=4_t=1.tiff"  "image_z=4_t=2.tiff"  "image_z=4_t=3.tiff"     "image_z=4_t=29.tiff"  "image_z=4_t=30.tiff"
 "image_z=5_t=1.tiff"  "image_z=5_t=2.tiff"  "image_z=5_t=3.tiff"     "image_z=5_t=29.tiff"  "image_z=5_t=30.tiff"

and then

julia> buf = load(first(filenames));

julia> img = load_series(load!, filenames, buf)

would create a 4-dimensional output. load! would ideally load directly into its first argument, but could be defined as

load!(dest, filename) = copyto!(dest, load(filename))

if needed.

MultifileArrays.select_series — Method

filenames = select_series(filepattern; dir=pwd())

Create a vector of filenames from filepattern. filepattern may be a string containing a * character (which is treated as a wildcard match), or a regular expression capturing a digit-substring. In either case, the */capture extracts an integer that determines file order.